+++ srunner/scenariomanager/scenarioatomics/atomic_trigger_conditions.py
@@ -120,6 +120,98 @@ class TriggerVelocity(AtomicBehavior):
         return new_status
 
 
+class AtStartCondition(AtomicBehavior):
+
+    """
+    This class contains a check if a named story element has started.
+
+    Important parameters:
+    - element_name: The story element's name attribute
+    - element_type: The element type [act,scene,maneuver,event,action]
+
+    The condition terminates with SUCCESS, when the named story element starts
+    """
+
+    def __init__(self, element_type, element_name):
+        """
+        Setup element details
+        """
+        super(AtStartCondition, self).__init__("AtStartCondition")
+        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
+        self._element_type = element_type
+        self._element_name = element_name
+        self._start_time = None
+        self._blackboard = py_trees.blackboard.Blackboard()
+
+    def initialise(self):
+        """
+        Initialize the start time of this condition
+        """
+        self._start_time = GameTime.get_time()
+        super(AtStartCondition, self).initialise()
+
+    def update(self):
+        """
+        Check if the specified story element has started since the beginning of the condition
+        """
+        new_status = py_trees.common.Status.RUNNING
+
+        blackboard_variable_name = "({}){}-{}".format(self._element_type.upper(), self._element_name, "START")
+        element_start_time = self._blackboard.get(blackboard_variable_name)
+        if element_start_time and element_start_time >= self._start_time:
+            new_status = py_trees.common.Status.SUCCESS
+
+        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
+
+        return new_status
+
+
+class AfterTerminationCondition(AtomicBehavior):
+
+    """
+    This class contains a check if a named story element has terminated.
+
+    Important parameters:
+    - element_name: The story element's name attribute
+    - element_type: The element type [act,scene,maneuver,event,action]
+
+    The condition terminates with SUCCESS, when the named story element ends
+    """
+
+    def __init__(self, element_type, element_name):
+        """
+        Setup element details
+        """
+        super(AfterTerminationCondition, self).__init__("AfterTerminationCondition")
+        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
+        self._element_type = element_type
+        self._element_name = element_name
+        self._start_time = None
+        self._blackboard = py_trees.blackboard.Blackboard()
+
+    def initialise(self):
+        """
+        Initialize the start time of this condition
+        """
+        self._start_time = GameTime.get_time()
+        super(AfterTerminationCondition, self).initialise()
+
+    def update(self):
+        """
+        Check if the specified story element has ended since the beginning of the condition
+        """
+        new_status = py_trees.common.Status.RUNNING
+
+        blackboard_variable_name = "({}){}-{}".format(self._element_type.upper(), self._element_name, "END")
+        element_start_time = self._blackboard.get(blackboard_variable_name)
+        if element_start_time and element_start_time >= self._start_time:
+            new_status = py_trees.common.Status.SUCCESS
+
+        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
+
+        return new_status
+
+
 class InTriggerRegion(AtomicBehavior):
 
     """